Refactor animal files
