+++ backend/src/main/java/de/neuefische/cgnjava222/zoomanagement/zoo/animal/Animal.java
@@ -3,5 +3,4 @@ package de.neuefische.cgnjava222.zoomanagement.zoo.animal;
 import org.springframework.data.annotation.Id;
 
 public record Animal(@Id String id, String name) {
-
 }
