+++ backend/src/main/java/de/neuefische/cgnjava222/zoomanagement/zoo/animal/AnimalRepo.java
@@ -5,5 +5,4 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface AnimalRepo extends MongoRepository<Animal, String> {
-
 }
