+++ backend/src/main/java/de/neuefische/cgnjava222/zoomanagement/zoo/animal/AnimalService.java
@@ -4,7 +4,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-
 @Service
 public class AnimalService {
