+++ backend/src/test/java/de/neuefische/cgnjava222/zoomanagement/zoo/animal/AnimalIntegrationTest.java
@@ -28,4 +28,5 @@ class AnimalIntegrationTest {
                         []
                         """));
     }
+    
 }
