+++ backend/src/test/java/de/neuefische/cgnjava222/zoomanagement/zoo/animal/AnimalServiceTest.java
@@ -8,9 +8,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-
 class AnimalServiceTest {
-
 
     @Test
     void getAnimals() {
@@ -21,18 +19,18 @@ class AnimalServiceTest {
         );
 
         AnimalRepo animalRepo = mock(AnimalRepo.class);
-        when(animalRepo.findAll())
-                .thenReturn(animals);
+        when(animalRepo.findAll()).thenReturn(animals);
 
         AnimalService animalService = new AnimalService(animalRepo);
 
-        List<Animal> result = animalService.getAnimals();
+        List<Animal> actualResult = animalService.getAnimals();
         List<Animal> expectedResult = List.of(
                 new Animal("1", "Eisbär"),
                 new Animal("2", "Tiger"),
                 new Animal("3", "Giraffe")
         );
 
-        assertThat(result).hasSameElementsAs(expectedResult);
+        assertThat(actualResult).hasSameElementsAs(expectedResult);
     }
+    
 }
