+++ frontend/src/components/animal/Animal.ts
